+++ Services/Interfaces/InterfacesLab2/IOrdinarySortingService2.cs
@@ -2,7 +2,7 @@
 
 namespace Laba.Services.Interfaces.InterfacesLab2
 {
-    public interface IOrdinarySortingService2: ISorting<int[], int>, ISwap<int>
+    public interface IOrdinarySortingService2: ISorting<int[], double>, ISwap<int>
     {
         public void PopulateBounds(int lower, int upper, int length);
     }
